backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.3.0
python-pptx>=0.6.23
scipy>=1.12.0
scikit-learn>=1.4.0
